dissector: restructure the dissectors for calling iso1 and iso2
Extended the dissector support to additional registered iso1 and
iso2 packet code. This is an interim commit to move some of
common utilities to v2gexi and test some display issues on
wireshark vs tshark for display issues.

The tshark will just call the dissectors with a tree pointer
that has been allocated, but wireshark does a two pass call
with a tree nil pointer and then a second pass with a valid
tree pointer - so, add the handshake for decode parts into
that.

Signed-off-by: Charles Hardin <charles.hardin@chargepoint.com>