dissector/v2giso1: first pass implementation for iso1
Parse and add the structs to the header fields and all of
the pointer tables. This is to enable testing and then verify
against available packet captures.

This should be testing against files since wireshark asserts
on invalid dissectors a live capture can end up losing state
when errors are encountered.

Signed-off-by: Charles Hardin <charles.hardin@chargepoint.com>