extern: patch the openv2g to increase the decoder limits
The XML schema definition doesn't impose max sizes on some of
the strings or arrays, so the openv2g code just picked arbitrary
values that are less than ideal. So, just increase these to get
a decode to work.

This is very brittle but it does get the pcap references provided
to successfully decode.

Fixes #25

Signed-off-by: Charles Hardin <charles.hardin@chargepoint.com>