feat(dissector): allow a preference to set the exi dissector to use
If the handshake isn't in the package trace then the dissector
cann't be determined from the initial exchange and so a setting
should support the exi decoder to use for the packet stream
instead.

There might be a better way to do this - but this is the a first
attempt.

Signed-off-by: Charles Hardin <charles.hardin@chargepoint.com>