chore(wireshark): bump build to use wireshark 4.0 infrastructure
The external plugin (standalone) infrastructure was changed in
4.0 to use the cmake configuration for the build information to
be useable. Unfortunately, this is not backwards compatible so
this is going to be one-way conversion to release.

This was exposed in the standlone builds on the CI/CD when the
git tree for the stable checkout of wireshark got bumped and caused
the builds to fail at that point in time.

Signed-off-by: Charles Hardin <charles.hardin@chargepoint.com>