exi_add_characters: escape non-printing and non-ASCII characters
If the character has a C string escape sequence, that sequence is used
as its representation. Otherwise, it is represented as an even number of
hexadigits inside \u{...}.

Since a pessimistic guess (12 * strlen) is used for the size of the
buffer needed for the converted c-string, alloca() is swapped out for
malloc()/free() to remove the risk of a stack overflow.

Fixes #44

Signed-off-by: Dennis Lambe Jr <dennis@sparkcharge.io>