doc: wireshark might not find the lua interpreter and not compile support
So, with wireshark 4.2 is using qt6 now and appears to be found fine
but the same issue is occurring with an older version of lua since
only 5.1 and 5.2 are supported by wireshark. Alter the documentation
to fetch the lua interpreter so the expected version is always used
regardless of what brew might have.

No functional change

Signed-off-by: Charles Hardin <charles.hardin@chargepoint.com>